Implement improved printscreen function
Implemented new print screen function that stores a jpg image of screen image in PrintScreen folder no matter what is sent to the local computer printer or print function. This now implements a permanence for printscreens instead of losing the image when print works or it does not work. It also stores full screen image as jpg.