Improve interface between SightLog and SR96 LOC forms
When a sight is pushed from SightLog form to SR96 LOC form, and key sight data fields are changed (date/time, lat/long, hs, ic, body/limb) and the user opts to update SightLog  a new sight will be created in SightLog. The user must then manage sights in SightLog to keep or delete any updated or created sights. IF key data fields are not changed and the user opts to update sight data in SightLog form, then all of the SR96 sight data is updated to SightLog.