Add required folder check to main menu
Add logic in main menu form to check that required celestial tools subfolders exist and if they do not create them to protect processing. EXCEPT HelpFiles is not created if it does not exist because it must be filled with files from winzip download install file. If HelpFiles does not exist then a warning message is flashed.